+++ kamu-cli/src/commands/sql_shell_command.rs
@@ -44,7 +44,7 @@ impl Command for SqlShellCommand {
     fn run(&mut self) -> Result<(), CLIError> {
         let sql_shell = SqlShellImpl::new(self.container_runtime.clone());
 
-        let spinner = if self.output_config.verbosity_level == 0 {
+        let spinner = if self.output_config.verbosity_level == 0 && !self.output_config.quiet {
             let mut pull_progress = PullImageProgress::new("container");
             sql_shell.ensure_images(&mut pull_progress);