+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.54.0] - 2021-08-15
+### Added
+- New `kamu inspect lineage` command that shows dependency tree of a dataset
+### Changed
+- The `kamu list depgraph` command was removed in favor of a specialized lineage output type: `kamu inspect lineage --all -o dot`
+
 ## [0.53.0] - 2021-08-13
 ### Fixed
 - Upgrading Spark engine with a workaround for upstream bug in Shapefile import